% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diagnosekobling}
\alias{diagnosekobling}
\title{Koblingsliste diagnosekoder}
\format{
\subsection{\code{diagnosekobling}}{

En tibble med 34 rader og 10 kolonner:
\describe{
\item{Kode}{ICD10-kode}
\item{diaggrupper_med}{diaggrupper_med-kode}
\item{diaggrupper_med_tekst}{Navn for diaggrupper_med-kode}
\item{diaggrupper_rem}{diaggrupper_rem-kode}
\item{diaggrupper_rem_tekst}{Navn for diaggrupper_rem-kode}
\item{rem_maal}{Remisjonsmål brukt for diaggrupper_rem-gruppe}
\item{diaggrupper_hoved}{diaggrupper_hoved-kode}
\item{diaggrupper_hoved_tekst}{Navn for diaggrupper_hoved-kode}
\item{perifer_aksial_diaggruppe}{perifer_aksial_diaggruppe-kode}
\item{perifer_aksial_diaggruppe_tekst}{Navn for perifer_aksial_diaggruppe-kode}
}
}
}
\source{
Laget av registeret og vedlikeholdt av statistiker ved fagsenteret
for medisinske kvalitetsregistre, region Vest.
}
\usage{
diagnosekobling
}
\description{
Koblingsliste for å knytte ICD10-diagnosekoder til diagnosegrupper spesifisert
av registeret. Inneholder fire ulike fordelinger av diagnoser som brukes blant
annet i årsrapport. Grupperingene er aktuelle i ulike sammenhenger hvor
enkelte diagnosegrupper skal presenteres samlet eller hver for seg.
}
\details{
Sist Oppdatert 2025-12-01.
}
\keyword{datasets}
