% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{fjern_uaktuelle_diagnoser}
\alias{fjern_uaktuelle_diagnoser}
\title{Fjern uaktuelle diagnoser}
\usage{
fjern_uaktuelle_diagnoser(diag)
}
\arguments{
\item{diag}{Diagnosedatasett fra NorArtritt.}
}
\value{
Returnerer diagnosedatasett hvor diagnosene
Artrose, Juvenil Idiopatisk Artritt (JIA), Kondrokalsinose,
Krystallartritter, Pyogen Artritt, Urinsyregikt, Septisk Artritt
og Polyartritt er filtrert bort.
}
\description{
Fjerner et utvalg uaktuelle diagnoser som er registrert i NorArtritt.
Disse er ikke av interesse for analyser, og hvis pasientene kun har en av
disse diagnosene oppfyller de ikke inklusjonskriteriene for registeret.
}
\examples{
d_full_Diagnoseskjema = tibble::tibble(
Navn = c("Artrose", "Kondrokalsinose", "Revmatoid Artritt"),
Kode = c("M130", "M139", "M058")
)

diagnosedata_filtrert = fjern_uaktuelle_diagnoser(d_full_Diagnoseskjema)
}
