% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{fjern_ugyldige_skjema}
\alias{fjern_ugyldige_skjema}
\title{Fjern ugyldige skjema}
\usage{
fjern_ugyldige_skjema(inkl, oppf, med, diag)
}
\arguments{
\item{inkl}{Datasett med inklusjonsskjema for NorArtritt.}

\item{oppf}{Datasett med oppfølgingsskjema for NorArtritt.}

\item{med}{Datasett med medisinskjema for NorArtritt.}

\item{diag}{Datasett med diagnoseskjema for NorArtritt.}
}
\value{
Returnerer filtrerte objekter for inndata med "_filtrert" som suffix.
}
\description{
Fjerner skjema i de ulike datasettene hvis pasienten mangler en nødvendig
registrering av et annet skjema.
For å inkluderes i analysene \emph{må} pasienten ha både inklusjons- og
diagnoseskjema.
Denne funksjonen fjerner skjema i følgende rekkefølge:
Diagnoseskjema for pasienter som mangler inklusjonsskjema.
Inklusjonsskjema for pasienter som mangler diagnoseskjema.
Oppfølgingsskjema for pasienter som mangler diagnoseskjema.
Medisineringsskjema for pasienter som mangler diagnoseskjema.
}
\examples{

d_full_Inklusjonskjema = tibble::tibble(
PasientGUID = c("1", "2", "3")
)

d_full_Oppfølgingskjema = tibble::tibble(
PasientGUID = c("1", "2", "3")
)

d_full_Medisineringskjema = tibble::tibble(
PasientGUID = c("1", "2", "3")
)

d_full_Diagnoseskjema = tibble::tibble(
PasientGUID = c("1", "2", "3"),
Navn = c("Artrose", "Kondrokalsinose", "Revmatoid Artritt"),
Kode = c("M130", "M139", "M058")
)

l_skjema = fjern_ugyldige_skjema(
  inkl = d_full_Inklusjonskjema,
  oppf = d_full_Oppfølgingskjema,
  med = d_full_Medisineringskjema,
  diag = d_full_Diagnoseskjema
)
}
