% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{fjerne_skjema_hjelpefunksjon}
\alias{fjerne_skjema_hjelpefunksjon}
\title{Hjelpefunksjon for fjerning av skjema}
\usage{
fjerne_skjema_hjelpefunksjon(d_hoved, d_motpart)
}
\arguments{
\item{d_hoved}{datasett med skjema som inneholder variabelen PasientGUID
d_motpart skal kontrolleres mot.}

\item{d_motpart}{datasett med skjema som eventuelt skal filtreres bort.
Må inneholde variabelen PasientGUID.}
}
\value{
Returnerer filtrert versjon av d_motpart, hvor ID'er som ikke finnes i
d_hoved er filtrert bort.
}
\description{
Hjelpefunksjon for fjerning av ugyldige skjema. Sjekker om det finnes
Pasient-ID i d_motpart som ikke finnes i d_hoved.
Brukes for eksempel for å fjerne inklusjonsskjema for pasienter som ikke har
diagnoseskjema.
}
