% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent_skjema.R
\name{hent_skjema}
\alias{hent_skjema}
\title{Hent skjema}
\usage{
hent_skjema(skjemanavn, registernavn = "data")
}
\arguments{
\item{skjemanavn}{Tekststreng med navn på skjemaet som skal hentes.}

\item{registernavn}{Tekststreng med navn på register. Standardverdi er "data".
Argumentet er med for å kunne kjøre tester på testdatabase.}
}
\value{
Tibble med alle data fra skjemaet med navn \code{skjemanavn}.
}
\description{
Henter alle data fra et skjema.
}
\details{
Fuksjonen bruker \code{\link[rapbase:loadRegData]{rapbase::loadRegData()}}
til å hente alle data fra skjemaet med navn \code{skjemanavn}:
\verb{SELECT * FROM skjemanavn}. Funksjonen er avhengig av
at database er koblet til for å fungere.
}
\examples{
\dontrun{
hent_skjema("Inklusjon")
}
}
