% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_kontroll}
\alias{ki_kontroll}
\title{Indikator for andel pasienter som er til kontroll innen 3 mnd}
\usage{
ki_kontroll(d_inkl_oppf, d_diag)
}
\arguments{
\item{d_inkl_oppf}{Sammenkoblet inklusjon- og oppfølgingsdatasett}

\item{d_diag}{Diagnosedatasett}
}
\value{
Returnerer en tibble med følgende variabler:
\itemize{
\item \strong{PasientGUID} Pasientidentifikator
\item \strong{ki_krit_teller} Indikator for om pasienten oppfyller
kriterier for teller. Kan ta verdiene TRUE, FALSE.
\item \strong{ki_krit_nevner} Indikator for om pasienten oppfyller
kriterier for nevner. Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler
}
}
\description{
Indikatoren viser andelen av pasientene diagnostisert med Revmatoid Artritt
som siste diagnose i 2014 eller senere, som får første oppfølging
innen 3 mnd (90 dager) av diagnosedato.
}
\details{
For pasienter som har inklusjonsskjema 7-90 dager etter diagnose regnes
inklusjonsskjema som første kontroll.
}
\examples{
d_inkl_oppf = tibble::tibble(
PasientGUID = c("1", "2", "3"),
InklusjonDato = as.Date(c("2021-01-01", "2022-01-01", "2023-01-01")),
DeathDate = c(NA, NA, NA),
dato_ktrl = as.Date(c("2021-01-10", "2022-01-10", "2023-01-10")),
Skjematype = c("Inklusjonskjema", "Inklusjonskjema", "Oppfølgingskjema")
)

d_diag = tibble::tibble(
PasientGUID = c("1", "2", "3"),
dato_diag = as.Date(c("2021-01-01", "2022-01-01", "2023-01-01")),
diaggrupper_med = c(1, 1, 1),
diaggrupper_hoved = c(1, 1, 1),
diag_stilt_aar = c(2021, 2022, 2023)
)

d_ki_kontroll = ki_kontroll(d_inkl_oppf, d_diag)
}
