% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_remisjon}
\alias{ki_remisjon}
\title{Andel RA-pasienter som oppnår remisjon 1 år etter diagnose.}
\usage{
ki_remisjon(d_diag, d_inkl_oppf, tidsrom_start = 180, tidsrom_slutt = 485)
}
\arguments{
\item{d_diag}{Diagnosedatasett.}

\item{d_inkl_oppf}{Sammenslått inklusjon- og oppfølgingsdatasett.}

\item{tidsrom_start}{Nedre grense av tidsrom for beregning av hva som er 1-års
kontroll. Antall dager etter diagnose.}

\item{tidsrom_slutt}{Øvre grense av tidsrom for beregning av hva som er 1-års
kontroll. Antall dager etter diagnose.}
}
\value{
Returnerer et datasett med følgende variabler:
\itemize{
\item \strong{PasientGUID} Pasientidentifikator.
\item \strong{ki_krit_teller} Indikator for om pasienten oppfyller
kriterier for teller. Kan ta verdiene TRUE, FALSE.
\item \strong{ki_krit_nevner} Indikator for om pasienten oppfyller
kriterier for nevner. Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler blir også med i utdata.
}
}
\description{
Indikatoren viser andelen pasienter med RA som har oppnådd remisjon 1 år
etter diagnosedato.
}
\details{
\code{tisdrom_start} og \code{tidsrom_slutt} indikerer start og slutt for hva som skal
regnes som 1-års kontroll i NorArtritt da det ikke er en fast kontroll etter
1 år. Pasienten må ha RA som siste diagnose, stilt i 2014 eller senere.
Pasienten må ha hatt diagnosen i minst 1 år ved datadumpdato, og ha vært
til kontroll i tidsrommet avgrenset med \code{tidsrom_start} og \code{tidsrom_slutt}.
I tillegg kan ingen av variablene for ACR-EULAR målingene være missing.
For å inkluderes i teller må også følgende krav være oppfylt for ACR-EULAR:
\itemize{
\item 1 eller færre ømme ledd.
\item 1 eller færre hovne ledd.
\item Crp <= 10.
\item Globalsykdomsaktivitet <= 10.
}
}
\examples{

d_diag = tibble::tibble(
PasientGUID = c("1", "2", "3"),
diaggrupper_med = c(1,1,1),
dato_diag = as.Date(c("2021-01-01", "2022-02-02", "2023-03-03")),
diag_stilt_aar = c(2021, 2022, 2023),
dager_diag_til_datadump = c(1826, 1429, 1035)
)

d_inkl_oppf = tibble::tibble(
PasientGUID = c("1", "2", "3"),
dato_ktrl = as.Date(c("2021-01-10", "2022-02-10", "2023-03-10")),
OmmeLeddAntall = c(1,2,3),
HovneLeddAntall = c(1,2,3),
Crp = c(1,2,3),
PasientGlobalSykdomsaktivitet = c(1,2,3),
DeathDate = as.Date(c(NA, NA, NA))
)

d_ki_remisjon = ki_remisjon(d_diag, d_inkl_oppf)
}
