% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{konverter_skjematype}
\alias{konverter_skjematype}
\title{Konverter skjematyper}
\usage{
konverter_skjematype(inkl, oppf)
}
\arguments{
\item{inkl}{datasett med alle inklusjonsskjema for NorArtritt. Må minimum
inneholde variablene PasientGUID, Skjematype, FormTypeId og SkjemaGUID.}

\item{oppf}{datasett med alle oppfølgingskskjema for NorArtritt. Må minimum
inneholde variablene PasientGUID, Skjematype, FormTypeId og SkjemaGUID.}
}
\value{
Returnerer en tibble med alle inklusjons- og oppfølgingsskjema samlet.
Dette vil videre filtreres og utbedres i påfølgende funksjoner før det til
slutt returneres som opprinnelige datasett hvor inklusjon og oppfølging er
separert.
}
\description{
\code{konverter_skjematype} håndterer manglende eller duplikate skjema for
pasienter i NorArtritt.
Alle pasienter \emph{skal} i teorien ha ett inklusjonsskjema, og et ubegrenset
antall oppfølgingsskjema. I virkeligheten er det dessverre ikke slik, og det
finnes pasienter med flere inklusjonsskjema og pasienter uten inklusjonsskjema.
}
\details{
Overflødige inklusjonsskjema konverteres til oppfølgingsskjema
(det tidligste inklusjonsskjema beholdes som inklusjonsskjema),
og hvis en pasient mangler inklusjonsskjema vil det tidligste oppfølgingsskjemaet
konverteres til inklusjonsskjema.
}
\examples{

d_full_Inklusjonskjema = tibble::tibble(
PasientGUID = c("1", "2", "3"),
FormTypeId = c(1,1,1),
FormDate = as.Date(c("2021-01-01", "2022-01-01", "2023-01-01")),
SkjemaGUID = c("d", "e", "f"),
Skjematype = c("Inklusjonskjema", "Inklusjonskjema", "Inklusjonskjema")
)

d_full_Oppfølgingsskjema = tibble::tibble(
PasientGUID = c("1", "2", "3"),
FormTypeId = c(2, 2, 2),
FormDate = as.Date(c("2021-01-10", "2022-01-10", "2023-01-10")),
SkjemaGUID = c("a", "b", "c"),
Skjematype = c("Oppfølgingskjema", "Oppfølgingskjema", "Oppfølgingskjema")
)

d_inkl_full_rettet = konverter_skjematype(
  inkl = d_full_Inklusjonskjema,
  oppf = d_full_Oppfølgingsskjema
)
}
