% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep-norartritt.R
\name{lag_filtrerte_objekter}
\alias{lag_filtrerte_objekter}
\title{Lag filtrerte objekter}
\usage{
lag_filtrerte_objekter(d_inkl, d_diag, d_med, d_oppf, datadump_dato = NULL)
}
\arguments{
\item{d_inkl}{Inklusjonsdatasett fra NorArtritt}

\item{d_diag}{Diagnosedatasett fra NorArtritt}

\item{d_med}{Medisindatasett fra NorArtritt}

\item{d_oppf}{Oppfølgingsskjema fra NorArtritt}

\item{datadump_dato}{Dato for datauttrekk}
}
\value{
Returnerer objektene d_med_vasket, d_diag_pers og d_diag_med til det
globale miljøet.
}
\description{
Lager filtrerte objekter som er nyttige for bruk i analyse til for eksempel
årsrapport.
}
\details{
Objektene som returneres er:
\itemize{
\item \strong{d_med_vasket} Inneholder alle medisinskjema for alle pasienter i
registeret som har en aktuell diagnose. Skjema uten startdato er fjernet, og
dupliserte skjema er tatt bort. For duplikater er skjema som har en
sluttdato beholdt.
\item \strong{d_diag_pers} Inneholder siste diagnose for hver pasient.
\item \strong{d_diag_med} Inneholder siste diagnose for hver pasient, i tillegg
til hele medisinhistorikken for pasientene. Her legges det også til
"Ingen medisin" for pasienter som ikke får medisinsk behandling.
}
}
