% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legemiddelkobling}
\alias{legemiddelkobling}
\title{Koblingsliste legemiddel}
\format{
\subsection{\code{legemiddelkobling}}{

En tibble med 98 rader og 2 kolonner:
\describe{
\item{legemiddel_kodebok_kode}{Komplett liste med alle legemiddel, inkludert
intern kode for legemiddel som importeres som kode 999}
\item{legemiddel_kodebok}{Navn for legemiddel}
}
}
}
\source{
Laget av registeret og vedlikeholdt av statistiker ved fagsenteret
for medisinske kvalitetsregistre, region Vest.
Sist oppdatert 2025-12-01.
}
\usage{
legemiddelkobling
}
\description{
Koblingsliste som brukes for 책 h책ndtere medisiner som registreres i GTI men
som ikke finnes i MRS. Disse blir importert med LegemiddelType = 999, men de
m책 skilles ut for 책 kunne brukes i analyse.
}
\keyword{datasets}
