% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_datovariabler}
\alias{legg_til_datovariabler}
\title{Legger til ekstra datovariabler}
\usage{
legg_til_datovariabler(d_inkl, d_oppf, d_med, d_diag)
}
\arguments{
\item{d_inkl}{inklusjonsskjema fra NorArtritt.}

\item{d_oppf}{Oppfølgingsskjema fra NorArtritt.}

\item{d_med}{Medisinskjema fra NorArtritt.}

\item{d_diag}{Diagnoseskjema fra NorArtritt.}
}
\value{
Returnerer en liste med de fire skjematypene i NorArtritt hvor ekstra
datovariabler er lagt til.
}
\description{
Legger til ekstra datovariabler til skjema i NorArtritt for å forenkle
analyser.
For inklusjon- og oppfølgingsskjema legges det til:
\itemize{
\item \strong{aar_ktrl} Årstall for inklusjon i registeret.
\item \strong{dato_ktrl} Dato for inklusjon i registeret.
For medisinskjema legges det til:
\item \strong{startaar} År for oppstart medisin.
\item \strong{sluttaar} År for avslutning av medisin.
\item \strong{StartDato} Dato for oppstart medisin.
\item \strong{SluttDato} Dato for avslutning av medisin.
}
}
\examples{
# leser inn data for norartritt:

d_inkl = tibble::tibble(
InklusjonDato = as.Date(c("2021-01-01", "2021-01-01"))
)

d_oppf = tibble::tibble(
FormDate = as.Date(c("2021-01-01", "2021-01-01"))
)

d_med = tibble::tibble(
StartDato = as.Date(c("2021-01-01", "2021-01-01")),
SluttDato = as.Date(c("2021-01-01", "2021-01-01"))
)

d_diag = tibble::tibble(
FormDate = as.Date(c("2021-01-01", "2021-01-01"))
)

l_skjema_med_ekstra_datovariabler = legg_til_datovariabler(
  d_inkl = d_inkl,
  d_oppf = d_oppf,
  d_med = d_med,
  d_diag = d_diag
)
}
