% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_diagnosegrupper}
\alias{legg_til_diagnosegrupper}
\title{Legg til diagnosegrupper}
\usage{
legg_til_diagnosegrupper(d)
}
\arguments{
\item{d}{Diagnosedatasett fra NorArtritt. Må inneholde variablene Kode og Navn.}
}
\value{
Returnerer inndata men med ekstra variabler med utbedret diagnoseinformasjon.
Nye variabler er:
\itemize{
\item \strong{diaggrupper_med} Heltallskode for de ulike diagnosegruppene. Det er
vanligvis denne inndelingen som brukes i NorArtritts analyser og rapporter.
Delt inn i 8 diagnosegrupper.
\item \strong{diaggrupper_med_tekst} Diagnosenavn tilhørende de ulike
heltallskodene i \code{diaggrupper_med}.
\item \strong{diaggrupper_rem} Heltallskode for inndeling i diagnosegrupper brukt
i sammenheng med beregning av remisjon. Disse måles på ulike måter, angitt i
variabelen \code{rem_maal}.
\item \strong{diaggrupper_rem_tekst} Diagnosenavn tilknyttet heltallskodene
angitt i \code{diaggrupper_rem}.
\item \strong{rem_maal} Navn for remisjonsmål knyttet heltallskoden i
\code{diaggrupper_rem}. Er enten DAS28-CRP eller ASDAS-CRP avhengig av diagnosen.
\item \strong{diaggrupper_hoved} Heltallskode for en alternativ diagnoseinndeling,
hvor diagnosene er delt i 5 grupper.
\item \strong{diaggrupper_hoved_tekst} Diagnosenavn tilknyttet kodene angitt i
\code{diaggrupper_hoved}.
\item \strong{perifer_aksial_diaggruppe} Heltallskode for en diagnoseinndeling
for perifer aksial artritt. Delt inn i 3 grupper.
\item \strong{perifer_aksial_diaggruppe_tekst} Gruppenavn for diagnosene delt etter
heltallskoden i \code{perifer_aksial_diaggruppe}.
}
}
\description{
Legger til ekstra kolonner for diagnoser til diagnosedata fra NorArtritt.
}
\details{
Funksjonen leser inn en diagnosefil fra NorArtritts kvalitetsserver.
Denne filen inneholder koblinger for alle diagnosene som er brukt i analyser
for NorArtritt, i tillegg til variabler for å gruppere diagnoser
etter ulike kriterier. For eksempel er kodene M058, M059 og M060 samlet til
en gruppe 'Revmatoid Artritt'.
'Leddsykdom' legges også i riktig grupper
selv om diagnosen ikke har en egen kode i registeret.
Ved endringer i hvilke diagnoser som inkluderes i
analyser er det i denne filen vi må oppdatere koblingene.
}
\examples{

d_diagnose = tibble::tibble(
Kode = c("M179", "M058", "M059"),
Navn = c("Diagnose", "RA", "Diagnose 2")
)

d_med_diagnosedata = legg_til_diagnosegrupper(d_diagnose)
}
