% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_medisinnavn}
\alias{legg_til_medisinnavn}
\title{Legg til medisinnavn}
\usage{
legg_til_medisinnavn(d)
}
\arguments{
\item{d}{Medisindatasett fra NorArtritt.}
}
\value{
Returnerer opprinnelig datasett med ekstra kolonner legemiddel_navn og
legemiddel_navn_kode som er vasket slik det er nevnt i beskrivelsen over.
}
\description{
Funksjon som grupperer legemiddel som har flere LegemiddelTyper og fjerner
'generisk' fra navn. Kobler også legemiddel som ikke importeres riktig ved
å bruke ATC kode og matche med eksisterende ATC-koder for kjente legemiddel.
}
\examples{

d_medisin_raa = tibble(ATC = c("L04AB04", "L04AB04", "L04AX03"),
                               LegemiddelType = c(46, 999, 59),
                               Legemiddel = c("AdalimumabGenerisk", "ImportertAnnet", "MethotrexateGenerisk"))

d_vasket_medisin = legg_til_medisinnavn(d_medisin_raa)
}
