% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_medisintype}
\alias{legg_til_medisintype}
\title{Legg til medisintype}
\usage{
legg_til_medisintype(d)
}
\arguments{
\item{d}{Medisindatasett fra norartritt.}
}
\value{
Returnerer opprinnelig datasett med følgende variabler koblet på
\itemize{
\item \strong{biokat} 1 hvis legemiddel er biologisk, 0 ellers.
\item \strong{dmard} 1 hvis legemiddel er dmard, 0 ellers.
\item \strong{csdmard} 1 hvis legemiddel er csdmard, 0 ellers.
\item \strong{tsdmard} 1 hvis legemiddel er tsdmard,  0 ellers.
\item \strong{bio_og_tsdmard} 1 hvis legemiddel er biologisk/tsdmard, 0 ellers.
}
}
\description{
Henter legemiddeltype-gruppering fra intern tabell. Denne tabellen inneholder
informasjon om hvilke av legemiddlene som er biologiske, dmards, csdmards,
tsdmards, og samlekategori biologiske_eller_tsdmard.

Kjøres som del av datavask-funksjon for NorArtritt.
}
\keyword{internal}
