% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medisinkobling}
\alias{medisinkobling}
\title{Koblingsliste medisiner}
\format{
\subsection{\code{medisinkobling}}{

En tibble med 93 rader og 15 kolonner:
\describe{
\item{LegemiddelType}{Kode for legemiddel i MRS}
\item{legemiddel_navn}{Prosessert navn for legemiddel}
\item{legemiddel_navn_kode}{Intern kode for legemiddel}
\item{biokat}{Indikator for om legemiddel er biologisk}
\item{dmard}{Indikator for om legemiddel er DMARD}
\item{csdmard}{Indikator for om legemiddel er csDMARD}
\item{tsdmard}{Indikator for om legemiddel er tsDMARD}
\item{bio_og_tsdmard}{Indikator for om legemiddel er Biologisk eller tsDMARD}
\item{legemiddel_gruppert}{Kode for legemiddel som skal grupperes (Samme virkestoff i ulike medisiner)}
\item{legemiddel_gruppert_navn}{Navn på gruppert legemiddel}
\item{Virkestoff}{Virkestoff i legemiddel}
\item{legemiddelnavn_i_kodebok}{Legemiddelnavn hentet fra kodebok}
\item{Kommentar}{Eventuelle kommentarer}
}
}
}
\source{
Laget av registeret og vedlikeholdt av statistiker ved fagsenteret
for medisinske kvalitetsregistre, region Vest.
Sist oppdatert 2025-12-01.
}
\usage{
medisinkobling
}
\description{
Koblingsliste for å knytte legemiddelType til legemiddel og virkestoff.
Inneholder også informasjon om hvilken type legemiddel hver medisin skal regnes
som.
}
\keyword{datasets}
