% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norartritt-package.R
\docType{package}
\name{norartritt-package}
\alias{norartritt}
\alias{norartritt-package}
\title{norartritt: Diverse funksjoner for registeret NorArtritt}
\description{
Diverse funksjoner for håndtering, analyse og rapportering av data for det medisinske kvalitetsregisteret NorArtritt (Norsk kvalitetsregister for artrittsykdommer).
}
\author{
\strong{Maintainer}: Endre Kvåle Evjen \email{endre.kvale.evjen@helse-bergen.no}

Authors:
\itemize{
  \item Endre Kvåle Evjen \email{endre.kvale.evjen@helse-bergen.no}
}

}
\keyword{internal}
