% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norartrittApp.R
\name{norartrittApp}
\alias{norartrittApp}
\title{Run the NorArtritt Shiny Application}
\usage{
norartrittApp()
}
\value{
An object representing the NorArtritt app
}
\description{
Run the NorArtritt Shiny Application
}
