% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{remisjon_totalt}
\alias{remisjon_totalt}
\title{Remisjon totalt}
\usage{
remisjon_totalt(d_diag, d_inkl_oppf)
}
\arguments{
\item{d_diag}{Diagnosedatasett}

\item{d_inkl_oppf}{Sammenkoblet inklusjon- og oppfølgingsdatasett}
}
\value{
Returnerer en tibble med følgende variabler:
\itemize{
\item \strong{PasientGUID} Pasientidentifikator.
\item \strong{ki_krit_teller} Indikator for om pasienten oppfyller
kriterier for teller. Kan ta verdiene TRUE, FALSE.
\item \strong{ki_krit_nevner} Indikator for om pasienten oppfyller
kriterier for nevner. Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler.
}
}
\description{
Regner ut total andel pasienter som har oppnådd remisjon.
}
\details{
Funksjonen beregner andelen av RA-pasientene som har oppnådd remisjon.
Ser kun på pasienter som har RA som siste diagnose.
}
\examples{
d_diag = tibble::tibble(
PasientGUID = c(1, 2, 3),
diaggrupper_med = c(1, 1, 1),
dato_diag = as.Date(c("2022-01-01", "2023-01-01", "2024-01-01")),
diag_stilt_aar = c(2022, 2023, 2024),
dager_diag_til_datadump = c(1461, 1096, 731),
)

d_inkl_oppf = tibble::tibble(
PasientGUID = c(1, 2, 3),
OmmeLeddAntall = c(10, 20, 25),
HovneLeddAntall = c(5, 8, 14),
Crp = c(10, 20, 30),
PasientGlobalSykdomsaktivitet = c(8, 12, 14),
)

d_remisjon_totalt = remisjon_totalt(d_diag, d_inkl_oppf)
}
