% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sykehuskobling}
\alias{sykehuskobling}
\title{Koblingsliste sykehusnavn}
\format{
\subsection{\code{sykehuskobling}}{

En tibble med 26 rader og 5 kolonner:
\describe{
\item{resh_id}{Resh-id for sykehus}
\item{sykehusnavn}{Leservennlig navn på sykehus}
\item{sykehus_kortnavn}{Leservennlig kortnavn for sykehus}
\item{sykehus_gruppe}{Kode for sykehusgruppe, nødvendig for å
samle avtalespesialister}
\item{sykehus_gruppe_navn}{Navn på gruppe}
}
}
}
\source{
Laget av registeret og vedlikeholdt av statistiker ved fagsenteret
for medisinske kvalitetsregistre, region Vest.
Sist oppdatert 2025-12-01.
}
\usage{
sykehuskobling
}
\description{
Koblingsliste for å lage leservennlige navn og kortnavn for sykehus og
avdelinger til bruk i rapporter og analyser.
}
\keyword{datasets}
