% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{velg_tidligste_inklusjondato}
\alias{velg_tidligste_inklusjondato}
\title{Velg tidligste inklusjondato}
\usage{
velg_tidligste_inklusjondato(d_inkl_oppf, pas_id = PasientGUID)
}
\arguments{
\item{d_inkl_oppf}{Sammenkoblet datasett som inneholder alle aktuelle
inklusjons- og oppfølgingsskjema for pasientgruppen.}

\item{pas_id}{Variabel som brukes som pasientidentifikator. Standard verdi
er \code{PasientGUID}, men for personidentifiserbare datadumper kan
\code{Fødselsnummer} brukes.}
}
\value{
Returnerer inndata med oppdatert \code{InklusjonDato} for alle pasienter.
}
\description{
Velger den tidligste inklusjonsdatoen for hver pasient oppgitt på
inklusjons- og oppfølgingsskjema i NorArtritt.
}
\details{
Inklusjons- og oppfølgingsskjema i NorArtritt har en variabel kalt
\code{InklusjonDato} som angir hvilken dato pasienten først ble inkludert i
registeret. Denne datoen skulle i realiteten vært identisk for alle
oppføringer for en pasient, men det er ikke tilfellet. Denne funksjonen
velger ut den tidligste registrerte \code{InklusjonDato} for hver pasient og
skriver over alle senere/manglende datoer i andre skjema.
Hvis en pasient aldri har fått registrert inklusjonsdato returneres NA for
denne pasienten.
}
\examples{

d_inkl_oppf = tibble::tibble(
PasientGUID = c("1", "2", "3"),
InklusjonDato = as.Date(c("2021-01-01", "2022-01-01", "2023-01-01"))
)

d_inkl_oppf_dato = velg_tidligste_inklusjondato(d_inkl_oppf)
}
