% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_asdas}
\alias{ki_asdas}
\title{Gjennomsnittlig ASDAS-CRP for spondyloartrittpasienter 1 år etter diagnose.}
\usage{
ki_asdas(d_diag, d_inkl_oppf, tidsrom_start = 180, tidsrom_slutt = 485)
}
\arguments{
\item{d_diag}{Diagnosedatasett.}

\item{d_inkl_oppf}{Sammenslått inklusjon- og oppfølgingsdatasett.}

\item{tidsrom_start}{Nedre grense av tidsrom for beregning av hva som er 1-års kontroll.
Antall dager etter diagnose.}

\item{tidsrom_slutt}{Øvre grense av tidsrom for beregning av hva som er 1-års kontroll.
Antall dager etter diagnose.}
}
\value{
Returnerer et datasett med følgende variabler:
\itemize{
\item \strong{PasientGUID} Pasientidentifikator.
\item \strong{ki_x} Numerisk verdi for ASDAS-CRP for pasienten.
\item \strong{ki_aktuell} Indikator for om pasienten oppfyller
kriterier for nevner. Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler blir også med i utdata.
}
}
\description{
Indikatoren viser gjennomsnittlig ASDAS-CRP 1 år etter diagnosedato
for pasienter med spondyloartritt
}
\details{
\code{tisdrom_start} og \code{tidsrom_slutt} indikerer start og slutt for hva som skal
regnes som 1-års kontroll i NorArtritt da det ikke er en fast kontroll etter
1 år. Pasienten må ha spondyloartritt som siste diagnose,
stilt i 2014 eller senere.
Pasienten må ha hatt diagnosen i minst 1 år ved datadumpdato, og ha vært
til kontroll i tidsrommet avgrenset med \code{tidsrom_start} og \code{tidsrom_slutt}.
I tillegg kan ikke variabelen for ASDAS-CRP være missing.

Dersom en pasient har hatt flere kontroller i det gyldige tidsintervallet
blir bare den kontrollen (som ikke har missing ASDAS-CRP-verdi)
som er nærmest 365 dager etter diagnose inkludert.
}
\examples{
# d_diag og d_inkl_oppf er diagnose og inklusjon/oppfølgingsdata fra NorArtritt.
d_ki_asdas = ki_asdas(d_diag, d_inkl_oppf)
}
