% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_kontroll_snitt}
\alias{ki_kontroll_snitt}
\title{Gjennomsnitt antall dager til første kontroll}
\usage{
ki_kontroll_snitt(d_ra_base, d_inkl_oppf)
}
\arguments{
\item{d_ra_base}{Forhåndsbehandlet datasett som inkluderer et uttrekk fra
registeret hvor vi har hentet ut pasienter som er aktuelle for de ulike
indikatorene for Revmatoid Artritt. Består av data fra inklusjonsskjema og
Diagnoseskjema.
Funksjonen \code{lag_ra_indikator_base()} kan brukes for å generere datasettet.}

\item{d_oppf}{Datasett med aktuelle oppfølgingsskjema.}
}
\value{
Returnerer et datasett med nødvendige koblingsnøkler.
Kolonnene \code{ki_x} og \code{ki_aktuell} inneholder antall dager fra diagnose til
første kontroll og indikator for om raden skal inkluderes eller ikke.
}
\description{
Indikatorfunksjon for å beregne gjennomsnittlig tid fra diagnose til første
oppfølging for pasienter diagnostisert med Revmatoid Artritt.
Funksjonen er ment å brukes på forhåndsbehandlede datasett hvor kobling
allerede er gjort. Dette kravet er satt for å flytte
kompleksitet fra indikatorfunksjoner til en egen filtreringsfunksjon.
}
\note{
Pasienter som ikke har vært til kontroll innen 365 dager fjernes fra
indikatoren. Pasienter som fikk diagnose for under 365 dager siden men som
enda ikke har vært til kontroll fjernes også.
}
\examples{
d_kontroll_snitt = ki_kontroll_snitt(d_ra_base, d_inkl_oppf)
}
