% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_medisinbruk}
\alias{ki_medisinbruk}
\title{Kvalitetsindikator for andel pasienter som bruker en medisin}
\usage{
ki_medisinbruk(d_diagnose, d_medisin, aarstall, legemiddel, diagnosekoder)
}
\arguments{
\item{d_diagnose}{Diagnosedatasett}

\item{d_medisin}{Medisindatasett}

\item{aarstall}{Årstallet indikatoren skal beregnes for}

\item{legemiddel}{En heltallsvektor med legemiddel_navn_kode for hvilke
medisiner som skal inkluderes.}

\item{diagnosekoder}{Hvilke diagnoser som skal inkluderes i beregningen.}
}
\value{
Returnerer en tibble med følgende variabler:
\itemize{
\item \strong{PasientGUID} PasientID
\item \strong{ki_krit_teller} Indikator for om pasienten oppfyller
kriterier for teller. Kan ta verdiene TRUE, FALSE.
\item \strong{ki_krit_nevner} Indikator for om pasienten oppfyller
kriterier for nevner. Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler blir også med i utdata.
}
}
\description{
Indikatoren viser andelen pasienter som bruker en gitt medisin i løpet av
et år.
}
\examples{
# d_diag og d_medisin er diagnosedata og medisindata
d_ki_medisinbruk = ki_medisinbruk(d_diagnose = d_diag, d_medisin = d_medisin, aarstall = 2020, legemiddel = 20, diagnosekoder = 1)
}
