% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_sykmod}
\alias{ki_sykmod}
\title{Kvalitetsindikator for tidlig oppstart av behandling}
\usage{
ki_sykmod(d_inklusjon, d_diagnose, d_medisin)
}
\arguments{
\item{d_inklusjon}{Inklusjonsdatasett}

\item{d_diagnose}{Diagnosedatasett}

\item{d_medisin}{Medisindatasett}
}
\value{
Returnerer en tibble med følgende variabler:
\itemize{
\item \strong{PasientGUID} PasientID
\item \strong{ki_krit_teller} Indikator for om pasienten oppfyller kriterier for teller.
Kan ta verdiene TRUE, FALSE.
\item \strong{ki_krit_nevner} Indikator for om pasienten oppfyller kriterier for nevner.
Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler blir også med i utdata.
}

Pasientene kan ha flere rader hvor ki_krit_teller har ulike besvarelse. For
eksempel kan en pasient ha startet på to medisiner innen 2 uker, men kun
en av medisinene er en sykdomsmodifiserende medisin. Utdata er da filtrert
slik at rader hvor ki_krit_nevner og ki_krit_teller er TRUE er prioritert.
}
\description{
Lager KI-datasett for om en pasient fikk tidlig oppstart av behandling.
}
\details{
Indikatoren viser andelen pasienter som får sykdomsmodifiserende medisin
innen 2 uker av diagnosedato.
Det er en del begrensninger for hvilke pasienter som skal inkluderes.
For å inkluderes i nevner må følgende kriterier være oppfylt:
\itemize{
\item Pasienten må ha Revmatoid Artritt som eneste registrerte diagnose.
\item Pasienten må være diagnostisert i 2014 eller senere.
\item Tid fra diagnose til inklusjon i registeret må være under eller lik 90 dager.
}

For å inkluderes i teller må de da i tillegg ha begynt på sykdomsmodifiserende
medisin innen 2 uker av diagnosedato.
}
\examples{
# d_inkl, d_diag og d_medisin er henholdsvis
# inklusjonsdata, diagnosedata og medisindata for NorArtitt.
d_ki_sykmod = ki_sykmod(d_inkl, d_diag, d_medisin)
}
