% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{ki_sykmod_snitt}
\alias{ki_sykmod_snitt}
\title{Gjennomsnitt antall dager til oppstart medisinbruk}
\usage{
ki_sykmod_snitt(d_ra_base, d_med)
}
\arguments{
\item{d_ra_base}{Forhåndsbehandlet datasett som inkluderer et uttrekk fra
registeret hvor vi har hentet ut pasienter som er aktuelle for de ulike
indikatorene for Revmatoid Artritt. Består av data fra inklusjonsskjema og
Diagnoseskjema.
Funksjonen \code{lag_ra_indikator_base()} kan brukes for å generere datasettet.}

\item{d_med}{Datasett med aktuelle medisinskjema.}
}
\value{
Returnerer et datasett med nødvendige koblingsnøkler.
Kolonnene \code{ki_x} og \code{ki_aktuell} inneholder antall dager fra diagnose til
oppstart medisinering og indikator for om raden skal inkluderes eller ikke.
}
\description{
Indikatorfunksjon for å beregne gjennomsnittlig tid fra diagnose til oppstart
medisinering med aktuell medisin. Funksjonen er ment å brukes på
forhåndsbehandlede datasett hvor kobling allerede er gjort.
Dette kravet er satt for å begrense kompleksitet i indikatorfunksjonen.
}
\note{
Pasienter som ikke har fått registrert medisinering innen 365 dager av diagnose
antas å ikke skulle ha medisinering, så disse vil derfor fjernes fra indikatoren.
Videre vil det være pasienter som ikke har startet medisinering og samtidig
ha fått diagnosen for mindre enn 365 dager siden ved tidspunktet for datauttrekk.
Disse pasientene (~ 5\% av de som fikk diagnose mindre enn 365 dager før
datadump-dato) fjernes fra indikator frem til enten det registreres
et medisinskjema (hvorpå de inkluderes) eller det går 365 dager fra diagnose
og de ekskluderes.
}
\examples{
d_ra_base = tibble::tibble(
PasientGUID = c("1", "2", "3"),
dato_diag = as.Date(c("2021-01-01", "2022-02-02", "2023-03-03")),
dager_diag_til_datadump = c(1826, 1429, 1035),
dager_diag_til_inklusjon = c(0, 1, 2),
DeathDate = c(NA, NA, NA),
)

d_med = tibble::tibble(
PasientGUID = c("1", "2", "3"),
StartDato = as.Date(c("2021-01-10", "2022-02-10", "2023-03-10")),
LegemiddelType = c(1,2,3),
dmard = c(1, 1, 1))

d_ki_sykmod = ki_sykmod_snitt(d_ra_base, d_med)
}
