% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{konverter_missing_til_na}
\alias{konverter_missing_til_na}
\title{Konverter missing til na}
\usage{
konverter_missing_til_na(d)
}
\arguments{
\item{d}{datasett med disse variablene. \code{inklusjonsskjema} eller
\code{oppfølgingsskjema}.}
}
\value{
Returnerer opprinnelig datasett hvor manglende besvarelser er erstattet med
\code{NA_real_}
}
\description{
Konverterer manglende besvarelser til NA for variablene
\code{BASDAI}, \code{Das28}, \code{Das283}, \code{Das28Crp}, \code{Das28Crp3},
\code{Cdai}, \code{Sdai}, \code{Asdas}, \code{DAPSA} og \code{Total}. Disse variablene er numeriske
verdier og manglende besvarelser er kodet som henholdsvis 0 for \code{BASDAI} og
-1 ellers i datasettet. Disse konverteres her til \code{NA_real_}.
\code{vask_data_norartritt()} kaller på denne funksjonen når
\code{lag_filtrerte_objekter()} kjøres.
}
\examples{
d_inkl = konverter_missing_til_na(d_inkl)
}
