% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{konverter_skjematype}
\alias{konverter_skjematype}
\title{Konverter skjematyper}
\usage{
konverter_skjematype(inkl, oppf)
}
\arguments{
\item{inkl}{datasett med alle inklusjonsskjema for NorArtritt. Må minimum
inneholde variablene PasientGUID, Skjematype, FormTypeId og SkjemaGUID.}

\item{oppf}{datasett med alle oppfølgingskskjema for NorArtritt. Må minimum
inneholde variablene PasientGUID, Skjematype, FormTypeId og SkjemaGUID.}
}
\value{
Returnerer en tibble med alle inklusjons- og oppfølgingsskjema samlet.
Dette vil videre filtreres og utbedres i påfølgende funksjoner før det til
slutt returneres som opprinnelige datasett hvor inklusjon og oppfølging er
separert.
}
\description{
\code{konverter_skjematype} håndterer manglende eller duplikate skjema for
pasienter i NorArtritt.
Alle pasienter \emph{skal} i teorien ha ett inklusjonsskjema, og et ubegrenset
antall oppfølgingsskjema. I virkeligheten er det dessverre ikke slik, og det
finnes pasienter med flere inklusjonsskjema og pasienter uten inklusjonsskjema.
}
\details{
Overflødige inklusjonsskjema konverteres til oppfølgingsskjema
(det tidligste inklusjonsskjema beholdes som inklusjonsskjema),
og hvis en pasient mangler inklusjonsskjema vil det tidligste oppfølgingsskjemaet
konverteres til inklusjonsskjema.
}
\examples{
# Leser inn data for NorArtritt
norartritt::les_data_norartritt()
d_inkl_full_rettet = konverter_skjematype(
  inkl = d_full_Inklusjonsskjema,
  oppf = d_full_Oppfølgingsskjema
)
}
