% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep-norartritt.R
\name{lag_ra_indikator_base}
\alias{lag_ra_indikator_base}
\title{lag base-data for ra-indikatorer}
\usage{
lag_ra_indikator_base(d_diag, d_inkl)
}
\arguments{
\item{d_diag}{Diagnosedatasett som skal brukes.}

\item{d_inkl}{Inklusjonsdatasett som skal brukes.}
}
\value{
En tibble med aktuelle variabler fra diagnoseskjema og inklusjonsskjema
filtrert for å kun inneholde pasienter som skal inkluderes i RA-indikatorer.
}
\description{
Lager et base-uttrekk for pasienter som skal inkluderes i indikatorer for
RA-pasienter.
Ytterligere skjema må kobles på for den enkelte indikator.
}
