% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_kriteriedatoer}
\alias{legg_til_kriteriedatoer}
\title{Velg tidligste registrerte dato for diagnose}
\usage{
legg_til_kriteriedatoer(d_inkl, d_diag, datadump_dato = NULL)
}
\arguments{
\item{d_inkl}{Inklusjonsskjema}

\item{d_diag}{Diagnoseskjema}

\item{datadump_dato}{Dato datadump er tatt ut.}
}
\value{
Returnerer opprinnelig diagnosedatasett med oppdatert dato_diag variabel
avhengig av diagnose og tilgjengelighet av kriteriedato som nevnt over.
Legger også til variablene \code{diag_stilt_aar} og \code{dager_diag_til_datadump}.
}
\description{
Det finnes flere ulike kilder for det som kan anses å være en diagnosedato.
Avhengig av diagnose er det ulike kriterier som skal oppfylles for at
diagnosen skal kunne stilles. Disse registreres på inklusjonsskjema, og vil
ikke nødvendigvis samsvare med dato fra diagnoseskjema.

Kriterie-variabler er som følger:
\verb{Revmatoid Artritt}: \code{OppfyltacrEularKriterierDato}
\verb{Aksial Spondyloartritt}: \code{AsasKriterierAksialDato}
\verb{Perifer Spondyloartritt}: \code{AsasKriterierPeriferDato}
\code{Psoriasisartritt}: \code{CasparKriterierDato}

Tidligste dato av overnevnte og Dato fra Diagnoseskjema brukes som diagnosedato.
}
\examples{
d_inkl = tibble(
PasientGUID = c("1", "2", "3", "4"),
AcrEularKlassifikasjonsKriterier = c(1,2,3,4),
OppfyltacrEularKriterier = c(1, 2, 2, 2),
OppfyltacrEularKriterierDato = as.Date("1991-01-01", NA, NA, NA),
AsasKriterierAksial = c(2, 1, 2, 2),
AsasKriterierAksialDato = c(NA, "2001-01-01", NA, NA),
AsasKriterierPerifer = c(2, 2, 1, 2),
AsasKriterierPeriferDato = c(NA, NA, "2002-01-01", NA),
CasparKriterier = c(2,2,2,1),
CasparKriterierDato = c(NA, NA, NA, "2020-01-01"),
)

datadump_dato = c("2026-01-01")

d_diag = tibble(
PasientGUID = c("1", "2", "3", "4"),
diaggrupper_med = c(1, 5, NA, 2),
Kode = c(NA, NA, "M138", NA),
dato_diag = c("2025-01-01", "2025-02-02", "2025-03-03"),
)

d_diag_kritdato = legg_til_kriteriedatoer(
d_inkl = d_inkl,
d_diag = d_diag,
dato_datadump = dato_datadump)
}
