% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_kriteriedatoer}
\alias{legg_til_kriteriedatoer}
\title{Velg tidligste registrerte dato for diagnose}
\usage{
legg_til_kriteriedatoer(d_inkl, d_diag)
}
\arguments{
\item{d_inkl}{Inklusjonsskjema}

\item{d_diag}{Diagnoseskjema}
}
\description{
Det finnes flere ulike kilder for det som kan anses å være en diagnosedato.
Avhengig av diagnose er det ulike kriterier som skal oppfylles for at
diagnosen skal kunne stilles. Disse registreres på inklusjonsskjema, og vil
ikke nødvendigvis samsvare med dato fra diagnoseskjema.

Kriterie-variabler er som følger:
\verb{Revmatoid Artritt}: \code{OppfyltacrEularKriterierDato}
\verb{Aksial Spondyloartritt}: \code{AsasKriterierAksialDato}
\verb{Perifer Spondyloartritt}: \code{AsasKriterierPeriferDato}
\code{Psoriasisartritt}: \code{CasparKriterierDato}

Tidligste dato av overnevnte og Dato fra Diagnoseskjema brukes som diagnosedato.
}
