% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{legg_til_sykehusnavn}
\alias{legg_til_sykehusnavn}
\title{Legg til sykehusnavn}
\usage{
legg_til_sykehusnavn(d)
}
\arguments{
\item{d}{Datasett fra NorArtritt. Må inneholde
kolonnen UnitId som tilsvarer RESH-id for de ulike avdelingene.}
}
\value{
Returnerer opprinnelig datasett i tillegg til ekstra informasjon hentet fra
sykehustabell på kvalitetsserver. Nye variabler er:
\itemize{
\item \strong{sykehusnavn} Fullt sykehusnavn.
Eksempel Haukeland universitetssykehus.
\item \strong{sykehus_kortnavn} Forkortet sykehusnavn. Eksempelvis er
Haukeland universitetssykehus forkortet til HUS.
\item \strong{sykehus_gruppe} Heltallskode for gruppering av sykehus. Primært
for å samle Avtalespesialister i en gruppe.
\item \strong{sykehus_gruppe_navn} Navn tilknyttet heltallskode
i \code{sykehus_gruppe}.
}
}
\description{
Kobler på sykehusnavn for bruk i analyser og rapporter.
}
\details{
NorArtritt har allerede variabler for sykehusnavn og sykehuskortnavn i
registeret, men disse variablene er mangelfulle. På grunn av det har vi
laget en egen tabell på kvalitetsserver hvor vi har oppdaterte og riktige
koblinger mellom UnitId og sykehusnavn. Det gjør det også enklere for oss
å gruppere sykehus etter ulike kriterier. Ved endringer i inkluderte sykehus
eller RESH-id'er er det denne tabellen som må oppdateres for å få riktige
navn i rapporter og analyser.
}
\examples{
# d er datasett fra NorArtritt som inneholder variabelen UnitId

d_med_sykehusnavn = legg_til_sykehusnavn(d)
}
