% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainnhenting.R
\name{les_data_norartritt}
\alias{les_data_norartritt}
\title{les data NorArtritt}
\usage{
les_data_norartritt(
  mappe_dd = NULL,
  dato = NULL,
  versjon = NULL,
  omgjevnad = .GlobalEnv
)
}
\arguments{
\item{mappe_dd}{Grunnmappe som datadumper skal hentes fra. Må ha undermapper
med datonavn (format: \code{ÅÅÅÅ-MM-DD}). Hvis \code{NULL} (standard), blir
grunnmappe_norartritt brukt.}

\item{dato}{Datoen for den aktuelle datadumpen (enten tekst på formatet \code{ÅÅÅÅ-MM-DD}
eller et \link[base:Dates]{base::Date}-objekt). Brukes for å velge riktig undermappe i
\code{mappe_dd}. Hvis \code{NULL} (standard), brukes siste dato i \code{mappe_dd}.}

\item{versjon}{Versjon av data som skal lastes ned. Standard er \code{MRS-PROD}.
Kan endres om det skal lastes ned data fra test eller QA-versjon av registeret.
Versjon er gjenspeilet i datafilnavn så det må stemme overens med navn på
filene som er lastet ned.}

\item{omgjevnad}{Miljøet som dataobjektene skal lagres til.
Standard er det globale miljøet.}
}
\description{
Leser inn data fra NorArtritt og lagrer de som objekt.
}
\details{
Leser inn standard datafiler for NorArtritt og lagrer som dataobjekt.
Foreløpig leses det kun inn hele datafiler, men validering skal legges til
når funksjonene for det er klare. Datafilene er ufiltrerte, så de inneholder
blant annet enkelte diagnoser som ikke er relevante for analyser.
Disse vil filtreres bort når valideringsfunksjonene er klare, og det vil
også returneres filtrerte og validerte datasett i denne funksjonen.
Det skal også legges inn støtte for de resterende datafilene som ikke brukes
per dags dato, men som kan tenkes å være nyttige i fremtiden.
}
\examples{
\dontrun{
# Les inn data (nyeste datadumper)
les_data_norartritt()
}
}
