% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer.R
\name{remisjon_totalt}
\alias{remisjon_totalt}
\title{Remisjon totalt}
\usage{
remisjon_totalt(d_diag, d_inkl_oppf)
}
\arguments{
\item{d_diag}{Diagnosedatasett}

\item{d_inkl_oppf}{Sammenkoblet inklusjon- og oppfølgingsdatasett}
}
\value{
Returnerer en tibble med følgende variabler:
\itemize{
\item \strong{PasientGUID} Pasientidentifikator.
\item \strong{ki_krit_teller} Indikator for om pasienten oppfyller
kriterier for teller. Kan ta verdiene TRUE, FALSE.
\item \strong{ki_krit_nevner} Indikator for om pasienten oppfyller
kriterier for nevner. Kan ta verdiene TRUE eller FALSE.
\item \strong{...} Eventuelle grupperingsvariabler.
}
}
\description{
Regner ut total andel pasienter som har oppnådd remisjon.
}
\details{
Funksjonen beregner andelen av RA-pasientene som har oppnådd remisjon.
Ser kun på pasienter som har RA som siste diagnose.
}
\examples{
# d_diag og d_inkl_oppf er diagnose og sammenslått
# inklusjon og oppfølgingsdatasett
d_remisjon_totalt = remisjon_totalt(d_diag, d_inkl_oppf)
}
