% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner-norartritt.R
\name{valider_legemiddeltype}
\alias{valider_legemiddeltype}
\title{Valider legemiddeltype}
\usage{
valider_legemiddeltype(mappe_dd)
}
\arguments{
\item{mappe_dd}{Filsti til hvor kodebok er lagret.}
}
\value{
Returnerer en feilmelding hvis legemiddeltyper i kodebok er endret.
}
\description{
Validerer legemiddeltype i medisinfil mot kodebok. Hvis HEMIT gjør endringer i
hvordan legemiddeltyper er kodet, eller legger til nye legemiddel i hovedlisten
fra de som per nå er kodet som LegemiddelType = 999 vil denne funksjonen
stanse innlesningen av medisinskjema og gi returnere en feilmelding.
}
\examples{
# mappe_dd er plassering for kodebok som skal sjekkes.

valider_legemiddeltype(mappe_dd)
}
