% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep-norartritt.R
\name{vask_data_norartritt}
\alias{vask_data_norartritt}
\title{vask data norartritt}
\usage{
vask_data_norartritt(d_inkl, d_oppf, d_diag, d_med)
}
\arguments{
\item{d_inkl}{Inklusjonsdatasett, vanligvis d_full_Inklusjonsskjema.}

\item{d_oppf}{Oppfølgingsdatasett, vanligvis d_full_Oppfølgingsskjema.}

\item{d_diag}{Diagnosedatasett, vanligvis d_full_Diagnoseskjema.}

\item{d_med}{Medisindatasett, vanligvis d_full_Medisineringskjema.}
}
\value{
Returnerer filtrerte og \emph{vaskede} datasett for ulike skjema i NorArtritt.

Basisskjema:
\itemize{
\item \strong{d_diag, d_med, d_inkl, d_oppf, d_inkl_oppf} \cr
Tilsvarer de fullstendige datasettene men med grunnleggende filtrering, slik
at blant annet ugyldige skjema er filtrert ut og datovariabler er kontrollert.
}

Koblede skjema:
\itemize{
\item \strong{d_med_vasket} \cr
Inneholder alle medisinskjema for alle pasienter i
registeret som har en aktuell diagnose. Skjema uten startdato er fjernet, og
dupliserte skjema er tatt bort. For duplikater er skjema som har en
sluttdato beholdt.
\item \strong{d_diag_pers} \cr
Inneholder siste diagnose for hver pasient.
\item \strong{d_diag_med} \cr
Inneholder siste diagnose for hver pasient, i tillegg
til hele medisinhistorikken for pasientene. Her legges det også til
"Ingen medisin" for pasienter som ikke får medisinsk behandling.
\item \strong{d_dodsdato} \cr
Inneholder oversikt over dødsdato, hvor tilgjengelig.
}
}
\description{
Funksjon som vasker data fra NorArtritt. Datadumper som lastes ned fra
registeret trenger preprossesering før det kan brukes i analyse.
Denne funksjonen gjør alt det nødvendige arbeidet med preprossesering
og lager nye objekter som brukes mye i analyse. Funksjonen vil normalt
sett ikke kalles direkte, men det kan være nyttig hvis en prøver å lese
inn et identifiserbart datasett med fødselsnummer, og ikke PasientGUID.
}
